XX-9632 - Cannot Delete the Empty Skill Groups when Mongodb is down

+++ sipXconfig/neoconf/src/org/sipfoundry/sipxconfig/openacd/OpenAcdContextImpl.java
@@ -467,7 +467,9 @@ public abstract class OpenAcdContextImpl extends SipxHibernateDaoSupport impleme
             }
         }
         getHibernateTemplate().deleteAll(groups);
-        m_provisioningContext.deleteObjects(skills);
+        if (!skills.isEmpty()) {
+            m_provisioningContext.deleteObjects(skills);
+        }
 
         return usedSkillGroups;
     }
